+++ packages/ui/core-components/src/lib/atoms/inputs/date-input/_DateInput.svelte
@@ -221,7 +221,7 @@
 				size="sm"
 				class={cn(
 					`flex justify-start ${range ? 'border-r rounded-r-none text-left' : 'text-center'} font-normal`,
-					!selectedDateInput && 'text-gray-400'
+					!selectedDateInput && 'text-base-content-muted'
 				)}
 				builders={[builder]}
 				disabled={!loaded}
@@ -311,11 +311,14 @@
 			bind:selected={selectedPreset}
 			disabled={!loaded}
 		>
-			<Select.Trigger class="h-8 w-40 rounded-l-none px-3 text-xs font-medium" sameWidth>
+			<Select.Trigger
+				class="h-8 w-40 rounded-l-none px-3 text-xs font-medium hover:bg-base-200 transition-colors"
+				sameWidth
+			>
 				{#if selectedPreset}
 					{selectedPreset.label}
 				{:else}
-					<span class="hidden sm:inline"> Select a Range </span>
+					<span class="hidden sm:inline text-base-content-muted"> Select a Range </span>
 					<span class="sm:hidden"> Range </span>
 				{/if}
 			</Select.Trigger>