+++ packages/ui/core-components/src/lib/unsorted/viz/core/Value.svelte
@@ -40,7 +40,7 @@
 			<EmptyChart {emptyMessage} {emptySet} {chartType} {isInitial} />
 		{/if}
 	</span>
-	<span slot="skeleton" class="text-gray-500">Loading...</span>
+	<span slot="skeleton" class="text-base-content-muted">Loading...</span>
 	<Value {...spreadProps} data={Query.isQuery(loaded) ? Array.from(loaded) : loaded}>
 		<slot />
 	</Value>