+++ sites/docs/components/DocTab.svelte
@@ -2,7 +2,9 @@
 	import { cubicInOut } from 'svelte/easing';
 	import { crossfade } from 'svelte/transition';
 
-	let activeTab = 'preview';
+	export let defaultTab = 'preview';
+
+	let activeTab = defaultTab;
 
 	const tabs = ['preview', 'code'];
 
@@ -32,7 +34,7 @@
 					</button>
 					{#if activeTab === tab}
 						<div
-							class="absolute bottom-0 left-0 w-full h-full rounded border bg-white shadow-sm z-0"
+							class="absolute bottom-0 left-0 w-full h-full rounded border border-base-300 bg-base-100 shadow-sm z-0"
 							in:send={{ key: 'trigger' }}
 							out:receive={{ key: 'trigger' }}
 						/>